requirements.txt
aiohttp
fake-useragent
google-genai
pydub
tenacity